fix: Rename create-icons to CommonJS extension for ES module compatibility
- Rename create-icons.js to create-icons.cjs
- Update build.sh to use the .cjs extension
- Fix ES module error in CI environment
